test: initial artifacts tests
