--- app/lib/artifact_ai/artifacts/artifacts.ex
@@ -1,3 +1,0 @@
-defmodule ArtifactAi.Artifacts do
-  @moduledoc false
-end
