--- app/test/artifact_ai/artifacts_test.exs
@@ -1,6 +1,0 @@
-defmodule ArtifactAi.ArtifactsTest do
-  use ArtifactAi.DataCase, async: true
-
-  describe "" do
-  end
-end
