--- app/test/artifact_ai/models/carts_test.exs
@@ -1,6 +1,0 @@
-defmodule ArtifactAi.CartsTest do
-  use ArtifactAi.DataCase, async: true
-
-  describe "carts" do
-  end
-end
