--- app/test/artifact_ai/models/offers_test.exs
@@ -1,6 +1,0 @@
-defmodule ArtifactAi.OffersTest do
-  use ArtifactAi.DataCase, async: true
-
-  describe "" do
-  end
-end
