--- app/test/artifact_ai/models/orders_test.exs
@@ -1,6 +1,0 @@
-defmodule ArtifactAi.OrdersTest do
-  use ArtifactAi.DataCase, async: true
-
-  describe "" do
-  end
-end
