--- app/test/artifact_ai/models/payments_test.exs
@@ -1,6 +1,0 @@
-defmodule ArtifactAi.PaymentsTest do
-  use ArtifactAi.DataCase, async: true
-
-  describe "" do
-  end
-end
