+++ app/test/support/fixtures/images_fixtures.ex
@@ -1,21 +1,22 @@
 defmodule ArtifactAi.ImagesFixtures do
   @moduledoc """
   This module defines test helpers for creating
-  entities via the `ArtifactAi.Images` context.
+  entities via the `ArtifactAi.Artifacts.Images` context.
   """
 
-  alias ArtifactAi.Artifacts.Images
+  alias ArtifactAi.Images
 
   @doc """
   Generate a prompt.
   """
-  def image_fixture(prompt, attrs \\ %{}) do
-    {:ok, image} =
+  def image_fixture(user, prompt, attrs \\ %{}) do
+    attrs =
       attrs
       |> Enum.into(%{
         url: "https://example.org"
       })
-      |> Images.create(prompt)
+
+    {:ok, image} = Images.create(user, prompt, attrs)
 
     image
   end
